+++ src/test/java/bumblebee/xchangepass/domain/user/service/UserLoginScenarioTest.java
@@ -9,8 +9,8 @@ import bumblebee.xchangepass.global.security.jwt.JwtProvider;
 import bumblebee.xchangepass.domain.user.login.LoginService;
 import bumblebee.xchangepass.domain.user.login.dto.request.LoginRequest;
 import bumblebee.xchangepass.domain.user.login.dto.response.LoginResponse;
-import bumblebee.xchangepass.domain.refresh.RefreshToken;
-import bumblebee.xchangepass.domain.refresh.RefreshTokenService;
+import bumblebee.xchangepass.domain.refresh.entity.RefreshToken;
+import bumblebee.xchangepass.domain.refresh.service.RefreshTokenService;
 import bumblebee.xchangepass.domain.refresh.dto.RefreshTokenResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
